feat: api 구조 역할별 분리
